Added the first trial for arduino code deployment
Changed the sensors that were used to  imitate IR sensors to the real IR sensors behavior. 
made a slight change in the ultrasonic calculation value to make the maximum range at 200cm. (needs testing)
made a slight change to the calculation of the distance to get more accurate results
checked the entire code and compiled it.
Next Steps: Test it in the Real Environment and adjust it as necessary .